Refactor module opt
