+++ ams/core/matprocessor.py
@@ -13,7 +13,7 @@ from andes.thirdparty.npfunc import safe_div
 from andes.shared import tqdm, tqdm_nb
 from andes.utils.misc import elapsed, is_notebook
 
-from ams.opt.omodel import Param
+from ams.opt import Param
 from ams.shared import pd, sps
 
 logger = logging.getLogger(__name__)
