+++ ams/opt/__init__.py
@@ -2,4 +2,11 @@
 Module for optimization modeling.
 """
 
-from ams.opt.omodel import Var, Constraint, Objective, OModel  # NOQA
+from ams.opt.optbase import OptzBase, ensure_symbols, ensure_mats_and_parsed  # NOQA
+from ams.opt.var import Var  # NOQA
+from ams.opt.exprcalc import ExpressionCalc  # NOQA
+from ams.opt.param import Param  # NOQA
+from ams.opt.constraint import Constraint  # NOQA
+from ams.opt.objective import Objective  # NOQA
+from ams.opt.expr import Expression  # NOQA
+from ams.opt.omodel import OModel  # NOQA
