+++ ams/opt/omodel.py
@@ -1,1094 +1,19 @@
 """
-Module for optimization modeling.
+Module for optimization OModel.
 """
 import logging
 
-from typing import Any, Optional, Union
+from typing import Any
 from collections import OrderedDict
-import re
 
-import numpy as np
-
-from andes.core.common import Config
 from andes.utils.misc import elapsed
 
 import cvxpy as cp
 
-from ams.utils import pretty_long_message
-from ams.shared import sps, _prefix, _max_length
+from ams.opt.optbase import ensure_symbols, ensure_mats_and_parsed
+
 
 logger = logging.getLogger(__name__)
-
-
-def ensure_symbols(func):
-    """
-    Decorator to ensure that symbols are generated before parsing.
-    If not, it runs self.rtn.syms.generate_symbols().
-
-    Designed to be used on the `parse` method of the optimization elements (`OptzBase`)
-    and optimization model (`OModel`), i.e., `Var`, `Param`, `Constraint`, `Objective`,
-    and `ExpressionCalc`.
-
-    Parsing before symbol generation can give wrong results. Ensure that symbols
-    are generated before calling the `parse` method.
-    """
-
-    def wrapper(self, *args, **kwargs):
-        if not self.rtn._syms:
-            logger.debug(f"<{self.rtn.class_name}> symbols are not generated yet. Generating now...")
-            self.rtn.syms.generate_symbols()
-        return func(self, *args, **kwargs)
-    return wrapper
-
-
-def ensure_mats_and_parsed(func):
-    """
-    Decorator to ensure that system matrices are built and the OModel is parsed
-    before evaluation. If not, it runs the necessary methods to initialize them.
-
-    Designed to be used on the `evaluate` method of optimization elements (`OptzBase`)
-    and the optimization model (`OModel`), i.e., `Var`, `Param`, `Constraint`, `Objective`,
-    and `ExpressionCalc`.
-
-    Evaluation before building matrices and parsing the OModel can lead to errors. Ensure that
-    system matrices are built and the OModel is parsed before calling the `evaluate` method.
-    """
-
-    def wrapper(self, *args, **kwargs):
-        try:
-            if not self.rtn.system.mats.initialized:
-                logger.debug("System matrices are not built yet. Building now...")
-                self.rtn.system.mats.build()
-            if isinstance(self, (OptzBase, Var, Param, Constraint, Objective)):
-                if not self.om.parsed:
-                    logger.debug("OModel is not parsed yet. Parsing now...")
-                    self.om.parse()
-            elif isinstance(self, OModel):
-                if not self.parsed:
-                    logger.debug("OModel is not parsed yet. Parsing now...")
-                    self.parse()
-        except Exception as e:
-            logger.error(f"Error during initialization or parsing: {e}")
-            raise
-        return func(self, *args, **kwargs)
-    return wrapper
-
-
-class OptzBase:
-    """
-    Base class for optimization elements.
-    Ensure that symbols are generated before calling the `parse` method. Parsing
-    before symbol generation can lead to incorrect results.
-
-    Parameters
-    ----------
-    name : str, optional
-        Name of the optimization element.
-    info : str, optional
-        Descriptive information about the optimization element.
-    unit : str, optional
-        Unit of measurement for the optimization element.
-
-    Attributes
-    ----------
-    rtn : ams.routines.Routine
-        The owner routine instance.
-    """
-
-    def __init__(self,
-                 name: Optional[str] = None,
-                 info: Optional[str] = None,
-                 unit: Optional[str] = None,
-                 ):
-        self.om = None
-        self.name = name
-        self.info = info
-        self.unit = unit
-        self.is_disabled = False
-        self.rtn = None
-        self.optz = None  # corresponding optimization element
-        self.code = None
-
-    @ensure_symbols
-    def parse(self):
-        """
-        Parse the object.
-        """
-        raise NotImplementedError
-
-    @ensure_mats_and_parsed
-    def evaluate(self):
-        """
-        Evaluate the object.
-        """
-        raise NotImplementedError
-
-    @property
-    def class_name(self):
-        """
-        Return the class name
-        """
-        return self.__class__.__name__
-
-    @property
-    def n(self):
-        """
-        Return the number of elements.
-        """
-        if self.owner is None:
-            return len(self.v)
-        else:
-            return self.owner.n
-
-    @property
-    def shape(self):
-        """
-        Return the shape.
-        """
-        try:
-            return self.om.__dict__[self.name].shape
-        except KeyError:
-            logger.warning('Shape info is not ready before initialization.')
-            return None
-
-    @property
-    def size(self):
-        """
-        Return the size.
-        """
-        if self.rtn.initialized:
-            return self.om.__dict__[self.name].size
-        else:
-            logger.warning(f'Routine <{self.rtn.class_name}> is not initialized yet.')
-            return None
-
-
-class ExpressionCalc(OptzBase):
-    """
-    Class for calculating expressions.
-
-    Note that `ExpressionCalc` is not a CVXPY expression, and it should not be used
-    in the optimization model.
-    It is used to calculate expression values **after** successful optimization.
-    """
-
-    def __init__(self,
-                 name: Optional[str] = None,
-                 info: Optional[str] = None,
-                 unit: Optional[str] = None,
-                 e_str: Optional[str] = None,
-                 model: Optional[Any] = None,
-                 src: Optional[str] = None,
-                 ):
-        OptzBase.__init__(self, name=name, info=info, unit=unit)
-        self.optz = None
-        self.e_str = e_str
-        self.code = None
-        self.model = model
-        self.owner = None
-        self.src = src
-        self.is_group = False
-
-    def get_idx(self):
-        if self.is_group:
-            return self.owner.get_idx()
-        elif self.owner is None:
-            logger.info(f'ExpressionCalc <{self.name}> has no owner.')
-            return None
-        else:
-            return self.owner.idx.v
-
-    @ensure_symbols
-    def parse(self):
-        """
-        Parse the Expression.
-        """
-        # parse the expression str
-        sub_map = self.om.rtn.syms.sub_map
-        code_expr = self.e_str
-        for pattern, replacement in sub_map.items():
-            try:
-                code_expr = re.sub(pattern, replacement, code_expr)
-            except Exception as e:
-                raise Exception(f"Error in parsing expr <{self.name}>.\n{e}")
-        # store the parsed expression str code
-        self.code = code_expr
-        msg = f" - ExpressionCalc <{self.name}>: {self.e_str}"
-        logger.debug(pretty_long_message(msg, _prefix, max_length=_max_length))
-        return True
-
-    @ensure_mats_and_parsed
-    def evaluate(self):
-        """
-        Evaluate the expression.
-        """
-        msg = f" - Expression <{self.name}>: {self.code}"
-        logger.debug(pretty_long_message(msg, _prefix, max_length=_max_length))
-        try:
-            local_vars = {'self': self, 'np': np, 'cp': cp}
-            self.optz = self._evaluate_expression(self.code, local_vars=local_vars)
-        except Exception as e:
-            raise Exception(f"Error in evaluating ExpressionCalc <{self.name}>.\n{e}")
-        return True
-
-    def _evaluate_expression(self, code, local_vars=None):
-        """
-        Helper method to evaluate the expression code.
-
-        Parameters
-        ----------
-        code : str
-            The code string representing the expression.
-
-        Returns
-        -------
-        cp.Expression
-            The evaluated cvxpy expression.
-        """
-        return eval(code, {}, local_vars)
-
-    @property
-    def v(self):
-        """
-        Return the CVXPY expression value.
-        """
-        if self.optz is None:
-            return None
-        else:
-            return self.optz.value
-
-    @property
-    def e(self):
-        """
-        Return the calculated expression value.
-        """
-        if self.code is None:
-            logger.info(f"ExpressionCalc <{self.name}> is not parsed yet.")
-            return None
-
-        val_map = self.om.rtn.syms.val_map
-        code = self.code
-        for pattern, replacement in val_map.items():
-            try:
-                code = re.sub(pattern, replacement, code)
-            except TypeError as e:
-                raise TypeError(e)
-
-        try:
-            logger.debug(pretty_long_message(f"Value code: {code}",
-                                             _prefix, max_length=_max_length))
-            local_vars = {'self': self, 'np': np, 'cp': cp, 'val_map': val_map}
-            return self._evaluate_expression(code, local_vars)
-        except Exception as e:
-            logger.error(f"Error in calculating expr <{self.name}>.\n{e}")
-            return None
-
-    def __repr__(self):
-        return f'{self.__class__.__name__}: {self.name}'
-
-
-class Expression(OptzBase):
-    """
-    Base class for expressions used in a routine.
-    """
-
-    def __init__(self,
-                 name: Optional[str] = None,
-                 info: Optional[str] = None,
-                 e_str: Optional[str] = None,
-                 ):
-        OptzBase.__init__(self, name=name, info=info)
-        self.e_str = e_str
-        self.optz = None
-        self.code = None
-
-    @ensure_symbols
-    def parse(self):
-        """
-        Parse the expression.
-
-        Returns
-        -------
-        bool
-            Returns True if the parsing is successful, False otherwise.
-        """
-        sub_map = self.om.rtn.syms.sub_map
-        code_expr = self.e_str
-        for pattern, replacement in sub_map.items():
-            try:
-                code_expr = re.sub(pattern, replacement, code_expr)
-            except Exception as e:
-                raise Exception(f"Error in parsing expr <{self.name}>.\n{e}")
-        self.code = code_expr
-        msg = f" - Expression <{self.name}>: {self.e_str}"
-        logger.debug(pretty_long_message(msg, _prefix, max_length=_max_length))
-        return True
-
-    @ensure_mats_and_parsed
-    def evaluate(self):
-        """
-        Evaluate the expression.
-
-        Returns
-        -------
-        bool
-            Returns True if the evaluation is successful, False otherwise.
-        """
-        msg = f" - Expression <{self.name}>: {self.code}"
-        logger.debug(pretty_long_message(msg, _prefix, max_length=_max_length))
-        try:
-            local_vars = {'self': self, 'np': np, 'cp': cp}
-            self.optz = self._evaluate_expression(self.code, local_vars=local_vars)
-        except Exception as e:
-            raise Exception(f"Error in evaluating Expression <{self.name}>.\n{e}")
-        return True
-
-    def _evaluate_expression(self, code, local_vars=None):
-        """
-        Helper method to evaluate the expression code.
-
-        Parameters
-        ----------
-        code : str
-            The code string representing the expression.
-
-        Returns
-        -------
-        cp.Expression
-            The evaluated cvxpy expression.
-        """
-        return eval(code, {}, local_vars)
-
-    @property
-    def v(self):
-        """
-        Return the CVXPY expression value.
-        """
-        if self.optz is None:
-            return None
-        else:
-            return self.optz.value
-
-    @property
-    def shape(self):
-        """
-        Return the shape.
-        """
-        try:
-            return self.om.__dict__[self.name].shape
-        except KeyError:
-            logger.warning('Shape info is not ready before initialization.')
-            return None
-
-    @property
-    def size(self):
-        """
-        Return the size.
-        """
-        if self.rtn.initialized:
-            return self.om.__dict__[self.name].size
-        else:
-            logger.warning(f'Routine <{self.rtn.class_name}> is not initialized yet.')
-            return None
-
-    @property
-    def e(self):
-        """
-        Return the calculated expression value.
-        """
-        if self.code is None:
-            logger.info(f"Expression <{self.name}> is not parsed yet.")
-            return None
-
-        val_map = self.om.rtn.syms.val_map
-        code = self.code
-        for pattern, replacement in val_map.items():
-            try:
-                code = re.sub(pattern, replacement, code)
-            except TypeError as e:
-                raise TypeError(e)
-
-        try:
-            logger.debug(pretty_long_message(f"Value code: {code}",
-                                             _prefix, max_length=_max_length))
-            local_vars = {'self': self, 'np': np, 'cp': cp, 'val_map': val_map}
-            return self._evaluate_expression(code, local_vars)
-        except Exception as e:
-            logger.error(f"Error in calculating expr <{self.name}>.\n{e}")
-            return None
-
-    def __repr__(self):
-        return f'{self.__class__.__name__}: {self.name}'
-
-
-class Param(OptzBase):
-    """
-    Base class for parameters used in a routine.
-
-    Parameters
-    ----------
-    no_parse: bool, optional
-        True to skip parsing the parameter.
-    nonneg: bool, optional
-        True to set the parameter as non-negative.
-    nonpos: bool, optional
-        True to set the parameter as non-positive.
-    cplx: bool, optional
-        True to set the parameter as complex, avoiding the use of `complex`.
-    imag: bool, optional
-        True to set the parameter as imaginary.
-    symmetric: bool, optional
-        True to set the parameter as symmetric.
-    diag: bool, optional
-        True to set the parameter as diagonal.
-    hermitian: bool, optional
-        True to set the parameter as hermitian.
-    boolean: bool, optional
-        True to set the parameter as boolean.
-    integer: bool, optional
-        True to set the parameter as integer.
-    pos: bool, optional
-        True to set the parameter as positive.
-    neg: bool, optional
-        True to set the parameter as negative.
-    sparse: bool, optional
-        True to set the parameter as sparse.
-    """
-
-    def __init__(self,
-                 name: Optional[str] = None,
-                 info: Optional[str] = None,
-                 unit: Optional[str] = None,
-                 no_parse: Optional[bool] = False,
-                 nonneg: Optional[bool] = False,
-                 nonpos: Optional[bool] = False,
-                 cplx: Optional[bool] = False,
-                 imag: Optional[bool] = False,
-                 symmetric: Optional[bool] = False,
-                 diag: Optional[bool] = False,
-                 hermitian: Optional[bool] = False,
-                 boolean: Optional[bool] = False,
-                 integer: Optional[bool] = False,
-                 pos: Optional[bool] = False,
-                 neg: Optional[bool] = False,
-                 sparse: Optional[bool] = False,
-                 ):
-        OptzBase.__init__(self, name=name, info=info, unit=unit)
-        self.no_parse = no_parse  # True to skip parsing the parameter
-        self.sparse = sparse
-
-        self.config = Config(name=self.class_name)  # `config` that can be exported
-
-        self.config.add(OrderedDict((('nonneg', nonneg),
-                                     ('nonpos', nonpos),
-                                     ('complex', cplx),
-                                     ('imag', imag),
-                                     ('symmetric', symmetric),
-                                     ('diag', diag),
-                                     ('hermitian', hermitian),
-                                     ('boolean', boolean),
-                                     ('integer', integer),
-                                     ('pos', pos),
-                                     ('neg', neg),
-                                     )))
-
-    @ensure_symbols
-    def parse(self):
-        """
-        Parse the parameter.
-
-        Returns
-        -------
-        bool
-            Returns True if the parsing is successful, False otherwise.
-        """
-        sub_map = self.om.rtn.syms.sub_map
-        code_param = "param(**config)"
-        for pattern, replacement, in sub_map.items():
-            try:
-                code_param = re.sub(pattern, replacement, code_param)
-            except Exception as e:
-                raise Exception(f"Error in parsing param <{self.name}>.\n{e}")
-        self.code = code_param
-        return True
-
-    @ensure_mats_and_parsed
-    def evaluate(self):
-        """
-        Evaluate the parameter.
-        """
-        if self.no_parse:
-            return True
-
-        config = self.config.as_dict()
-        try:
-            msg = f"Parameter <{self.name}> is set as sparse, "
-            msg += "but the value is not sparse."
-            if self.sparse:
-                self.v = sps.csr_matrix(self.v)
-
-            # Create the cvxpy.Parameter object
-            if isinstance(self.v, np.ndarray):
-                self.optz = cp.Parameter(shape=self.v.shape, **config)
-            else:
-                self.optz = cp.Parameter(**config)
-            self.optz.value = self.v
-        except ValueError:
-            msg = f"Parameter <{self.name}> has non-numeric value, "
-            msg += "set `no_parse=True`."
-            logger.warning(msg)
-            self.no_parse = True
-            return True
-        except Exception as e:
-            raise Exception(f"Error in evaluating Param <{self.name}>.\n{e}")
-        return True
-
-    def update(self):
-        """
-        Update the Parameter value.
-        """
-        # NOTE: skip no_parse parameters
-        if self.optz is None:
-            return None
-        self.optz.value = self.v
-        return True
-
-    def __repr__(self):
-        return f'{self.__class__.__name__}: {self.name}'
-
-
-class Var(OptzBase):
-    """
-    Base class for variables used in a routine.
-
-    When `horizon` is provided, the variable will be expanded to a matrix,
-    where rows are indexed by the source variable index and columns are
-    indexed by the horizon index.
-
-    Parameters
-    ----------
-    info : str, optional
-        Descriptive information
-    unit : str, optional
-        Unit
-    tex_name : str
-        LaTeX-formatted variable symbol. Defaults to the value of ``name``.
-    name : str, optional
-        Variable name. One should typically assigning the name directly because
-        it will be automatically assigned by the model. The value of ``name``
-        will be the symbol name to be used in expressions.
-    src : str, optional
-        Source variable name. Defaults to the value of ``name``.
-    model : str, optional
-        Name of the owner model or group.
-    horizon : ams.routines.RParam, optional
-        Horizon idx.
-    nonneg : bool, optional
-        Non-negative variable
-    nonpos : bool, optional
-        Non-positive variable
-    cplx : bool, optional
-        Complex variable
-    imag : bool, optional
-        Imaginary variable
-    symmetric : bool, optional
-        Symmetric variable
-    diag : bool, optional
-        Diagonal variable
-    psd : bool, optional
-        Positive semi-definite variable
-    nsd : bool, optional
-        Negative semi-definite variable
-    hermitian : bool, optional
-        Hermitian variable
-    boolean : bool, optional
-        Boolean variable
-    integer : bool, optional
-        Integer variable
-    pos : bool, optional
-        Positive variable
-    neg : bool, optional
-        Negative variable
-
-    Attributes
-    ----------
-    a : np.ndarray
-        Variable address.
-    _v : np.ndarray
-        Local-storage of the variable value.
-    rtn : ams.routines.Routine
-        The owner routine instance.
-    """
-
-    def __init__(self,
-                 name: Optional[str] = None,
-                 tex_name: Optional[str] = None,
-                 info: Optional[str] = None,
-                 src: Optional[str] = None,
-                 unit: Optional[str] = None,
-                 model: Optional[str] = None,
-                 shape: Optional[Union[tuple, int]] = None,
-                 v0: Optional[str] = None,
-                 horizon=None,
-                 nonneg: Optional[bool] = False,
-                 nonpos: Optional[bool] = False,
-                 cplx: Optional[bool] = False,
-                 imag: Optional[bool] = False,
-                 symmetric: Optional[bool] = False,
-                 diag: Optional[bool] = False,
-                 psd: Optional[bool] = False,
-                 nsd: Optional[bool] = False,
-                 hermitian: Optional[bool] = False,
-                 boolean: Optional[bool] = False,
-                 integer: Optional[bool] = False,
-                 pos: Optional[bool] = False,
-                 neg: Optional[bool] = False,
-                 ):
-        self.name = name
-        self.info = info
-        self.unit = unit
-
-        self.tex_name = tex_name if tex_name else name
-        # variable internal index inside a model (assigned in run time)
-        self.id = None
-        OptzBase.__init__(self, name=name, info=info, unit=unit)
-        self.src = src
-        self.is_group = False
-        self.model = model  # indicate if this variable is a group variable
-        self.owner = None  # instance of the owner model or group
-        self.v0 = v0
-        self.horizon = horizon
-        self._shape = shape
-        self._v = None
-        self.a: np.ndarray = np.array([], dtype=int)
-
-        self.config = Config(name=self.class_name)  # `config` that can be exported
-
-        self.config.add(OrderedDict((('nonneg', nonneg),
-                                     ('nonpos', nonpos),
-                                     ('complex', cplx),
-                                     ('imag', imag),
-                                     ('symmetric', symmetric),
-                                     ('diag', diag),
-                                     ('psd', psd),
-                                     ('nsd', nsd),
-                                     ('hermitian', hermitian),
-                                     ('boolean', boolean),
-                                     ('integer', integer),
-                                     ('pos', pos),
-                                     ('neg', neg),
-                                     )))
-
-    @property
-    def v(self):
-        """
-        Return the CVXPY variable value.
-        """
-        if self.optz is None:
-            return None
-        if self.optz.value is None:
-            try:
-                shape = self.optz.shape
-                return np.zeros(shape)
-            except AttributeError:
-                return None
-        else:
-            return self.optz.value
-
-    @v.setter
-    def v(self, value):
-        if self.optz is None:
-            logger.info(f"Variable <{self.name}> is not initialized yet.")
-        else:
-            self.optz.value = value
-
-    def get_idx(self):
-        if self.is_group:
-            return self.owner.get_idx()
-        elif self.owner is None:
-            logger.info(f'Variable <{self.name}> has no owner.')
-            return None
-        else:
-            return self.owner.idx.v
-
-    @ensure_symbols
-    def parse(self):
-        """
-        Parse the variable.
-        """
-        sub_map = self.om.rtn.syms.sub_map
-        # NOTE: number of rows is the size of the source variable
-        if self.owner is not None:
-            nr = self.owner.n
-            nc = 0
-            if self.horizon:
-                # NOTE: numer of columns is the horizon if exists
-                nc = int(self.horizon.n)
-                shape = (nr, nc)
-            else:
-                shape = (nr,)
-        elif isinstance(self._shape, int):
-            shape = (self._shape,)
-            nr = shape
-            nc = 0
-        elif isinstance(self._shape, tuple):
-            shape = self._shape
-            nr = shape[0]
-            nc = shape[1] if len(shape) > 1 else 0
-        else:
-            raise ValueError(f"Invalid shape {self._shape}.")
-        code_var = f"var({shape}, **config)"
-        logger.debug(f" - Var <{self.name}>: {self.code}")
-        for pattern, replacement, in sub_map.items():
-            try:
-                code_var = re.sub(pattern, replacement, code_var)
-            except Exception as e:
-                raise Exception(f"Error in parsing var <{self.name}>.\n{e}")
-        self.code = code_var
-        return True
-
-    @ensure_mats_and_parsed
-    def evaluate(self):
-        """
-        Evaluate the variable.
-
-        Returns
-        -------
-        bool
-            Returns True if the evaluation is successful, False otherwise.
-        """
-        # NOTE: in CVXPY, Config only allow lower case letters
-        config = {}  # used in `self.code`
-        for k, v in self.config.as_dict().items():
-            if k == 'psd':
-                config['PSD'] = v
-            elif k == 'nsd':
-                config['NSD'] = v
-            elif k == 'bool':
-                config['boolean'] = v
-            else:
-                config[k] = v
-        msg = f" - Var <{self.name}>: {self.code}"
-        logger.debug(pretty_long_message(msg, _prefix, max_length=_max_length))
-        try:
-            local_vars = {'self': self, 'config': config, 'cp': cp}
-            self.optz = eval(self.code, {}, local_vars)
-        except Exception as e:
-            raise Exception(f"Error in evaluating Var <{self.name}>.\n{e}")
-        return True
-
-    def __repr__(self):
-        return f'{self.__class__.__name__}: {self.owner.__class__.__name__}.{self.name}'
-
-
-class Constraint(OptzBase):
-    """
-    Base class for constraints.
-
-    This class is used as a template for defining constraints. Each
-    instance of this class represents a single constraint.
-
-    Parameters
-    ----------
-    name : str, optional
-        A user-defined name for the constraint.
-    e_str : str, optional
-        A mathematical expression representing the constraint.
-    info : str, optional
-        Additional informational text about the constraint.
-    is_eq : str, optional
-        Flag indicating if the constraint is an equality constraint. False indicates
-        an inequality constraint in the form of `<= 0`.
-
-    Attributes
-    ----------
-    is_disabled : bool
-        Flag indicating if the constraint is disabled, False by default.
-    rtn : ams.routines.Routine
-        The owner routine instance.
-    is_disabled : bool, optional
-        Flag indicating if the constraint is disabled, False by default.
-    dual : float, optional
-        The dual value of the constraint.
-    code : str, optional
-        The code string for the constraint
-    """
-
-    def __init__(self,
-                 name: Optional[str] = None,
-                 e_str: Optional[str] = None,
-                 info: Optional[str] = None,
-                 is_eq: Optional[bool] = False,
-                 ):
-        OptzBase.__init__(self, name=name, info=info)
-        self.e_str = e_str
-        self.is_eq = is_eq
-        self.is_disabled = False
-        self.dual = None
-        self.code = None
-
-    @ensure_symbols
-    def parse(self):
-        """
-        Parse the constraint.
-        """
-        # parse the expression str
-        sub_map = self.om.rtn.syms.sub_map
-        code_constr = self.e_str
-        for pattern, replacement in sub_map.items():
-            try:
-                code_constr = re.sub(pattern, replacement, code_constr)
-            except TypeError as e:
-                raise TypeError(f"Error in parsing constr <{self.name}>.\n{e}")
-        # parse the constraint type
-        code_constr += " == 0" if self.is_eq else " <= 0"
-        # store the parsed expression str code
-        self.code = code_constr
-        msg = f" - Constr <{self.name}>: {self.e_str}"
-        logger.debug(pretty_long_message(msg, _prefix, max_length=_max_length))
-        return True
-
-    def _evaluate_expression(self, code, local_vars=None):
-        """
-        Helper method to evaluate the expression code.
-
-        Parameters
-        ----------
-        code : str
-            The code string representing the expression.
-
-        Returns
-        -------
-        cp.Expression
-            The evaluated cvxpy expression.
-        """
-        return eval(code, {}, local_vars)
-
-    @ensure_mats_and_parsed
-    def evaluate(self):
-        """
-        Evaluate the constraint.
-        """
-        msg = f" - Constr <{self.name}>: {self.code}"
-        logger.debug(pretty_long_message(msg, _prefix, max_length=_max_length))
-        try:
-            local_vars = {'self': self, 'cp': cp, 'sub_map': self.om.rtn.syms.val_map}
-            self.optz = self._evaluate_expression(self.code, local_vars=local_vars)
-        except Exception as e:
-            raise Exception(f"Error in evaluating Constraint <{self.name}>.\n{e}")
-
-    def __repr__(self):
-        enabled = 'OFF' if self.is_disabled else 'ON'
-        out = f"{self.class_name}: {self.name} [{enabled}]"
-        return out
-
-    @property
-    def e(self):
-        """
-        Return the calculated constraint LHS value.
-        Note that `v` should be used primarily as it is obtained
-        from the solver directly.
-
-        `e` is for debugging purpose. For a successfully solved problem,
-        `e` should equal to `v`. However, when a problem is infeasible
-        or unbounded, `e` can be used to check the constraint LHS value.
-        """
-        if self.code is None:
-            logger.info(f"Constraint <{self.name}> is not parsed yet.")
-            return None
-
-        val_map = self.om.rtn.syms.val_map
-        code = self.code
-        for pattern, replacement in val_map.items():
-            try:
-                code = re.sub(pattern, replacement, code)
-            except TypeError as e:
-                raise TypeError(e)
-
-        try:
-            logger.debug(pretty_long_message(f"Value code: {code}",
-                                             _prefix, max_length=_max_length))
-            local_vars = {'self': self, 'np': np, 'cp': cp, 'val_map': val_map}
-            return self._evaluate_expression(code, local_vars)
-        except Exception as e:
-            logger.error(f"Error in calculating constr <{self.name}>.\n{e}")
-            return None
-
-    @property
-    def v(self):
-        """
-        Return the CVXPY constraint LHS value.
-        """
-        if self.optz is None:
-            return None
-        if self.optz._expr.value is None:
-            try:
-                shape = self._expr.shape
-                return np.zeros(shape)
-            except AttributeError:
-                return None
-        else:
-            return self.optz._expr.value
-
-    @v.setter
-    def v(self, value):
-        raise AttributeError("Cannot set the value of the constraint.")
-
-
-class Objective(OptzBase):
-    """
-    Base class for objective functions.
-
-    This class serves as a template for defining objective functions. Each
-    instance of this class represents a single objective function that can
-    be minimized or maximized depending on the sense ('min' or 'max').
-
-    Parameters
-    ----------
-    name : str, optional
-        A user-defined name for the objective function.
-    e_str : str, optional
-        A mathematical expression representing the objective function.
-    info : str, optional
-        Additional informational text about the objective function.
-    sense : str, optional
-        The sense of the objective function, default to 'min'.
-        `min` for minimization and `max` for maximization.
-
-    Attributes
-    ----------
-    v : NoneType
-        The value of the objective function. It needs to be set through
-        computation.
-    rtn : ams.routines.Routine
-        The owner routine instance.
-    code : str
-        The code string for the objective function.
-    """
-
-    def __init__(self,
-                 name: Optional[str] = None,
-                 e_str: Optional[str] = None,
-                 info: Optional[str] = None,
-                 unit: Optional[str] = None,
-                 sense: Optional[str] = 'min'):
-        OptzBase.__init__(self, name=name, info=info, unit=unit)
-        self.e_str = e_str
-        self.sense = sense
-        self.code = None
-
-    @property
-    def e(self):
-        """
-        Return the calculated objective value.
-
-        Note that `v` should be used primarily as it is obtained
-        from the solver directly.
-
-        `e` is for debugging purpose. For a successfully solved problem,
-        `e` should equal to `v`. However, when a problem is infeasible
-        or unbounded, `e` can be used to check the objective value.
-        """
-        if self.code is None:
-            logger.info(f"Objective <{self.name}> is not parsed yet.")
-            return None
-
-        val_map = self.om.rtn.syms.val_map
-        code = self.code
-        for pattern, replacement in val_map.items():
-            try:
-                code = re.sub(pattern, replacement, code)
-            except TypeError as e:
-                logger.error(f"Error in parsing value for obj <{self.name}>.")
-                raise e
-
-        try:
-            logger.debug(pretty_long_message(f"Value code: {code}",
-                                             _prefix, max_length=_max_length))
-            local_vars = {'self': self, 'np': np, 'cp': cp, 'val_map': val_map}
-            return self._evaluate_expression(code, local_vars)
-        except Exception as e:
-            logger.error(f"Error in calculating obj <{self.name}>.\n{e}")
-            return None
-
-    @property
-    def v(self):
-        """
-        Return the CVXPY objective value.
-        """
-        if self.optz is None:
-            return None
-        else:
-            return self.optz.value
-
-    @v.setter
-    def v(self, value):
-        raise AttributeError("Cannot set the value of the objective function.")
-
-    @ensure_symbols
-    def parse(self):
-        """
-        Parse the objective function.
-
-        Returns
-        -------
-        bool
-            Returns True if the parsing is successful, False otherwise.
-        """
-        # parse the expression str
-        sub_map = self.om.rtn.syms.sub_map
-        code_obj = self.e_str
-        for pattern, replacement, in sub_map.items():
-            try:
-                code_obj = re.sub(pattern, replacement, code_obj)
-            except Exception as e:
-                raise Exception(f"Error in parsing obj <{self.name}>.\n{e}")
-        # store the parsed expression str code
-        self.code = code_obj
-        if self.sense not in ['min', 'max']:
-            raise ValueError(f'Objective sense {self.sense} is not supported.')
-        sense = 'cp.Minimize' if self.sense == 'min' else 'cp.Maximize'
-        self.code = f"{sense}({code_obj})"
-        msg = f" - Objective <{self.name}>: {self.code}"
-        logger.debug(pretty_long_message(msg, _prefix, max_length=_max_length))
-        return True
-
-    @ensure_mats_and_parsed
-    def evaluate(self):
-        """
-        Evaluate the objective function.
-
-        Returns
-        -------
-        bool
-            Returns True if the evaluation is successful, False otherwise.
-        """
-        logger.debug(f" - Objective <{self.name}>: {self.e_str}")
-        try:
-            local_vars = {'self': self, 'cp': cp}
-            self.optz = self._evaluate_expression(self.code, local_vars=local_vars)
-        except Exception as e:
-            raise Exception(f"Error in evaluating Objective <{self.name}>.\n{e}")
-        return True
-
-    def _evaluate_expression(self, code, local_vars=None):
-        """
-        Helper method to evaluate the expression code.
-
-        Parameters
-        ----------
-        code : str
-            The code string representing the expression.
-
-        Returns
-        -------
-        cp.Expression
-            The evaluated cvxpy expression.
-        """
-        return eval(code, {}, local_vars)
-
-    def __repr__(self):
-        return f"{self.class_name}: {self.name} [{self.sense.upper()}]"
 
 
 class OModel:
