+++ ams/routines/dopf.py
@@ -7,7 +7,7 @@ from ams.core.param import RParam
 
 from ams.routines.dcopf import DCOPF
 
-from ams.opt.omodel import Var, Constraint, Objective
+from ams.opt import Var, Constraint, Objective
 
 
 class DOPF(DCOPF):
